+++ src/widgets/settings/notefoldersettingswidget.h
@@ -77,6 +77,7 @@ class NoteFolderSettingsWidget : public QWidget {
                                                             const QString &text);
     QString generatePathFromCurrentNoteFolderRemotePathItem(QTreeWidgetItem *item);
     void updateSubfolderVisibility();
+    void updateCloudConnectionEnabledState();
     void updateNoteFolderListActiveState();
     void populateSubfolderTree();
     void populateSubfolderTreeFromDir(QTreeWidgetItem *parentItem, const QString &path,